+++ docfx.sln
@@ -64,8 +64,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DocAsCode.Common.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DocAsCode.Glob.Tests", "test\Microsoft.DocAsCode.Glob.Tests\Microsoft.DocAsCode.Glob.Tests.csproj", "{A95C5BAE-1D0D-44FD-A984-7CBF055A62E2}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DocAsCode.Build.MemberLevelManagedReference", "plugins\Microsoft.DocAsCode.Build.MemberLevelManagedReference\Microsoft.DocAsCode.Build.MemberLevelManagedReference.csproj", "{2746DA6E-D310-4F07-B41B-929BB3D6E14B}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DocAsCode.Plugins.Tests", "test\Microsoft.DocAsCode.Plugins.Tests\Microsoft.DocAsCode.Plugins.Tests.csproj", "{B3EEB5FE-CF39-4CC7-9650-5709E71D2BA9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DocAsCode.Build.ManagedReference.Tests", "test\Microsoft.DocAsCode.Build.ManagedReference.Tests\Microsoft.DocAsCode.Build.ManagedReference.Tests.csproj", "{958DF90F-0528-4C50-9AC2-E86C60971B7D}"
@@ -382,18 +380,6 @@ Global
 		{A95C5BAE-1D0D-44FD-A984-7CBF055A62E2}.Release|x64.Build.0 = Release|Any CPU
 		{A95C5BAE-1D0D-44FD-A984-7CBF055A62E2}.Release|x86.ActiveCfg = Release|Any CPU
 		{A95C5BAE-1D0D-44FD-A984-7CBF055A62E2}.Release|x86.Build.0 = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|x64.Build.0 = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Debug|x86.Build.0 = Debug|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|x64.ActiveCfg = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|x64.Build.0 = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|x86.ActiveCfg = Release|Any CPU
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B}.Release|x86.Build.0 = Release|Any CPU
 		{B3EEB5FE-CF39-4CC7-9650-5709E71D2BA9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B3EEB5FE-CF39-4CC7-9650-5709E71D2BA9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B3EEB5FE-CF39-4CC7-9650-5709E71D2BA9}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -722,7 +708,6 @@ Global
 		{05A49D39-97F0-46AF-9EDC-76D1E19FE3AC} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
 		{B2EEB5FE-CF39-4CC7-9650-5709E71D2BA9} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
 		{A95C5BAE-1D0D-44FD-A984-7CBF055A62E2} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
-		{2746DA6E-D310-4F07-B41B-929BB3D6E14B} = {DF2BA00B-1742-4F84-A469-08D7F019498E}
 		{B3EEB5FE-CF39-4CC7-9650-5709E71D2BA9} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
 		{958DF90F-0528-4C50-9AC2-E86C60971B7D} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
 		{958DF90F-0528-4C53-9AC2-E86C60971B7D} = {926A0726-B806-4215-82EF-AF8E22D0FACF}
