+++ src/Microsoft.DocAsCode.Dotnet/DotnetApiCatalog.cs
@@ -123,7 +123,8 @@ public static class DotnetApiCatalog
             CodeSourceBasePath = configModel?.CodeSourceBasePath,
             DisableDefaultFilter = configModel?.DisableDefaultFilter ?? false,
             NoRestore = configModel?.NoRestore ?? false,
-            NamespaceLayout = configModel?.NamespaceLayout ?? NamespaceLayout.Flattened,
+            NamespaceLayout = configModel?.NamespaceLayout ?? default,
+            MemberLayout = configModel?.MemberLayout ?? default,
             AllowCompilationErrors = configModel?.AllowCompilationErrors ?? false,
             Files = expandedFiles.Items.SelectMany(s => s.Files).ToList(),
             References = expandedReferences?.Items.SelectMany(s => s.Files).ToList(),
