+++ src/Microsoft.DocAsCode.Dotnet/ExtractMetadata/ExtractMetadataConfig.cs
@@ -27,6 +27,8 @@ internal class ExtractMetadataConfig
 
     public NamespaceLayout NamespaceLayout { get; init; }
 
+    public MemberLayout MemberLayout { get; init; }
+
     public Dictionary<string, string> MSBuildProperties { get; init; }
 
     public bool AllowCompilationErrors { get; init; }
