+++ templates/default/ManagedReference.html.primary.js
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-var mrefCommon = require('./ManagedReference.common.js');
+var common = require('./ManagedReference.common.js');
 var extension = require('./ManagedReference.extension.js');
 var overwrite = require('./ManagedReference.overwrite.js');
 
@@ -15,8 +15,8 @@ exports.transform = function (model) {
     model = extension.preTransform(model);
   }
 
-  if (mrefCommon && mrefCommon.transform) {
-    model = mrefCommon.transform(model);
+  if (common && common.transform) {
+    model = common.transform(model);
   }
   if (model.type.toLowerCase() === "enum") {
     model.isClass = false;
@@ -25,6 +25,10 @@ exports.transform = function (model) {
   model._disableToc = model._disableToc || !model._tocPath || (model._navPath === model._tocPath);
 
   if (extension && extension.postTransform) {
+    if (model._splitReference) {
+      model = postTransformMemberPage(model);
+    }
+
     model = extension.postTransform(model);
   }
 
@@ -35,8 +39,27 @@ exports.getOptions = function (model) {
   if (overwrite && overwrite.getOptions) {
     return overwrite.getOptions(model);
   }
+  var ignoreChildrenBookmarks = model._splitReference && model.type && common.getCategory(model.type) === 'ns';
 
   return {
-    "bookmarks": mrefCommon.getBookmarks(model)
+    "bookmarks": common.getBookmarks(model, ignoreChildrenBookmarks)
   };
+}
+
+function postTransformMemberPage(model) {
+  var type = model.type.toLowerCase();
+  var category = common.getCategory(type);
+  if (category == 'class') {
+      var typePropertyName = common.getTypePropertyName(type);
+      if (typePropertyName) {
+          model[typePropertyName] = true;
+      }
+      if (model.children && model.children.length > 0) {
+          model.isCollection = true;
+          common.groupChildren(model, 'class');
+      } else {
+          model.isItem = true;
+      }
+  }
+  return model;
 }
