+++ templates/modern/partials/class.header.tmpl.partial
@@ -13,9 +13,11 @@
 <div class="markdown summary">{{{summary}}}</div>
 <div class="markdown conceptual">{{{conceptual}}}</div>
 
+{{#syntax.content.0.value}}
 <div class="codewrapper">
   <pre><code class="lang-{{_lang}} hljs">{{syntax.content.0.value}}</code></pre>
 </div>
+{{/syntax.content.0.value}}
 
 {{#syntax.parameters.0}}
 <h4 class="section">{{__global.parameters}}</h4>
@@ -32,7 +34,7 @@
 {{#syntax.return}}
 <h4 class="section">{{__global.returns}}</h4>
 <dl class="parameters">
-  <dt>type.specName.0.value</dt>
+  <dt>{{{type.specName.0.value}}}</dt>
   <dd>{{{description}}}</dd>
 </dl>
 {{/syntax.return}}
