+++ templates/modern/partials/collection.tmpl.partial
@@ -1,7 +1,3 @@
 {{!Copyright (c) Microsoft. All rights reserved. Licensed under the MIT license. See LICENSE file in the project root for full license information.}}
-{{#isCollection}}
-{{>partials/collection}}
-{{/isCollection}}
-{{#isItem}}
-{{>partials/item}}
-{{/isItem}}
+
+{{>partials/class}}
