+++ templates/modern/partials/customMREFContent.tmpl.partial
